Move datastore types to their own file
