+++ src/app/queries/types/index.ts
@@ -4,4 +4,5 @@ export * from './mappings.types';
 export * from './plans.types';
 export * from './providers.types';
 export * from './networks.types';
+export * from './datastores.types';
 export * from './storageClasses.types';
